Moduleify ElaDetails and add smoke test
